[Feat] 유저 관련 API 구현 완료

See merge request s13-webmobile1-sub1/S13P11A404!30

+++ frontend/BE/src/main/java/com/a404/duckonback/entity/Artist.java
@@ -17,7 +17,7 @@ public class Artist {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer artistId;
+    private Long artistId;
 
     @Column(name = "name_en", nullable = false, length = 100)
     private String nameEn;
